[GraphTrainer] Extract remove-noop graph passes into dedicated module
Move remove_detach_pass, remove_identity_view_pass,
remove_identity_slice_pass, and _IDENTITY_VIEW_TARGETS from passes.py
into a new remove_noop_passes.py module. The passes are re-exported from
passes.py so all existing callsites (including tests) continue to work
without changes.

This reduces the size of the monolithic passes.py and groups related
no-op removal passes together for better discoverability.
